experiments/pixel_face_lab_ai/requirements.txt
torch>=2.2.0
diffusers==0.30.3
transformers==4.46.3
peft>=0.12.0
accelerate>=0.33.0
safetensors>=0.4.0
Pillow>=10.0.0,<12.0.0
numpy>=1.26.0
opencv-python>=4.10.0
